+++ cypress/integration/pagination.spec.js
@@ -213,7 +213,7 @@ context("Pagination", () => {
       cy.server();
       cy.route({
         method: "GET",
-        url: "/jobs/search?full_time=false&description=&location1=Chicago",
+        url: "/jobs/search?full_time=false&contract=false&description=&location1=Chicago",
         status: 200,
         response: jobsJson,
       });
@@ -224,7 +224,7 @@ context("Pagination", () => {
 
       cy.paginationSelect1(childList);
 
-      cy.get(":nth-child(3) > [data-cy=checkmark]").click();
+      cy.get(":nth-child(4) > [data-cy=checkmark]").click();
       cy.get("#search-submit").click();
 
       cy.get('[data-cy="job-container"]').then(($jobs) => {