+++ src/client/components/OptionsPanel/OptionsPanel-styled.tsx
@@ -1,6 +1,7 @@
 import styled from "styled-components";
 
 const OptionsPanelContainer = styled.div`
+  padding: 10px;
   width: 25%;
 
   @media only screen and (max-width: 800px) {
@@ -9,6 +10,10 @@ const OptionsPanelContainer = styled.div`
 
   @media only screen and (max-width: 600px) {
     width: 100%;
+  }
+
+  button {
+    margin-top: 10px;
   }
 `;