+++ src/client/modals/HiddenJobs/HiddenJobs.tsx
@@ -9,54 +9,56 @@ import { getHiddenJobsDetails } from "../../redux/thunks";
 import { RootState, Job } from "../../types";
 
 export interface HiddenJobsProps {
+  currentPage: number;
   handleGetHiddenJobsDetails: () => void;
-  hiddenJobsCurrentPage: number;
+  hiddenJobs: string[];
   hiddenJobsDetails: Job[];
-  hiddenJobsTotalPages: number;
 }
 
 const HiddenJobs: React.SFC<HiddenJobsProps> = (props: HiddenJobsProps) => {
   const {
+    currentPage,
     handleGetHiddenJobsDetails,
-    hiddenJobsCurrentPage,
+    hiddenJobs,
     hiddenJobsDetails,
-    hiddenJobsTotalPages,
   } = props;
 
-  const jobsOnPage =
-    hiddenJobsDetails &&
-    hiddenJobsDetails.slice(
-      hiddenJobsCurrentPage * 5 - 5,
-      hiddenJobsCurrentPage * 5
-    );
+  const [jobsOnPage, setJobsOnPage] = React.useState([]);
 
   React.useEffect((): void => {
-    handleGetHiddenJobsDetails();
-  }, []);
+    if (hiddenJobsDetails) {
+      setJobsOnPage(
+        hiddenJobsDetails.slice(currentPage * 5 - 5, currentPage * 5)
+      );
+    }
+  }, [hiddenJobsDetails]);
+
+  React.useEffect(() => {
+    if (hiddenJobs.length > 0) {
+      handleGetHiddenJobsDetails();
+    } else {
+      setJobsOnPage([]);
+    }
+  }, [hiddenJobs]);
+
+  const pages = Math.ceil(hiddenJobs.length / 5);
 
   return (
     <div>
       {jobsOnPage &&
         jobsOnPage.map((job: Job) => <JobCard job={job} key={job.id} />)}
       {jobsOnPage.length > 0 && (
-        <Pagination
-          currentPage={hiddenJobsCurrentPage}
-          totalPages={hiddenJobsTotalPages}
-        />
+        <Pagination currentPage={currentPage} totalPages={pages} />
       )}
-      {jobsOnPage.length === 0 && (
-        <div id="no-results">
-          No results. Please modify your search and try again.
-        </div>
-      )}
+      {jobsOnPage.length === 0 && <div id="no-results">No results.</div>}
     </div>
   );
 };
 
 const mapStateToProps = (state: RootState) => ({
-  hiddenJobsCurrentPage: state.user.hiddenJobsCurrentPage,
+  currentPage: state.application.currentPage,
+  hiddenJobs: state.user.hiddenJobs,
   hiddenJobsDetails: state.user.hiddenJobsDetails,
-  hiddenJobsTotalPages: state.user.hiddenJobsTotalPages,
 });
 
 const mapDispatchToProps = (dispatch) => ({