+++ src/client/modals/ResetPassword/ResetPassword-styled.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const ActionsContainer = styled.div`
+  display: flex;
+
+  button:nth-child(1) {
+    margin-right: 15px;
+  }
+`;
+
+export { ActionsContainer };